Removed dateToSeconds to dataUtils
